+++ src/routers/Search/dashboard/src/components/AppendArtifactDialog/TagsStep.svelte
@@ -4,32 +4,15 @@
 -->
 <script lang="ts">
   import Paper, { Subtitle, Content } from "@smui/paper";
-  import CircularProgress from "@smui/circular-progress";
   import FileButton from "../FileButton.svelte";
-  import SchemaForm, { ValidationError, type JSONSchema7, type UISchema } from "svelte-jsonschema-form";
-  import { deepMerge } from "../../../../Utils";
+  import SchemaForm from "../SchemaForm.svelte";
 
   export let data: any;
   export let disabled = false
 
-  let schema = fetchSchema();
   let tagsFile: FileList | null = null;
-  let schemaForm: SchemaForm;
-  let schemaError: string | Error | ValidationError | null = null;
-  let uischema = { ":ui:": { "collapse": "unrequired" }} as UISchema;
 
-  $: updateUischema(disabled);
   $: mergeTagsFile(tagsFile?.[0]);
-
-  async function fetchSchema(): Promise<JSONSchema7> {
-    const url = "../schema.json";    
-    const response = await fetch(url);
-    if (response.ok) {
-      return response.json();
-    } else {
-      throw new Error(await response.text());
-    }
-  }
 
   async function mergeTagsFile(file: File | undefined) {
     if (file) {
@@ -40,21 +23,6 @@
       reader.readAsText(file);
     }
   }
-
-  function updateUischema(disabled: boolean) {
-    uischema = deepMerge(uischema, <UISchema>{ ":ui:": { "readonly": disabled }})
-  }
-
-  function handleSchemaFormError(event: CustomEvent<Error | ValidationError>) {
-    console.error(event.detail);
-    if (event.detail instanceof ValidationError) {
-      schemaError = event.detail;
-    }
-    else {
-      schemaError = "Invalid JSON Schema. Please check the configuration file.";
-    }
-  }
-
 </script>
 
 <Paper variant="unelevated" class="tags-step">
@@ -64,14 +32,7 @@
   </Subtitle>
 
   <Content>
-    {#await schema}
-      <CircularProgress indeterminate />
-      <p>Loading schema...</p>
-    {:then schema}
-      <SchemaForm {schema} {uischema} bind:data={data} bind:this={schemaForm} on:error={handleSchemaFormError} />
-    {:catch error}
-      <div class="error">ERROR: {error.message}</div>
-    {/await}
+    <SchemaForm bind:data={data} />
   </Content>
 </Paper>
